requirements.txt
alabaster==0.7.12
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.11.0
decorator==5.1.0
docutils==0.17.1
flake8==4.0.1
idna==3.3
imagesize==1.2.0
iniconfig==1.1.1
Jinja2==3.0.2
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
nptyping==1.4.4
numpy==1.21.3
numpydoc==1.1.0
packaging==21.0
pandas==1.3.4
Pillow==8.4.0
pluggy==1.0.0
py==1.10.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
pymed==0.8.9
pyparsing==3.0.1
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
retry==0.9.2
scikit-learn==1.0.1
scipy==1.7.1
six==1.16.0
sklearn==0.0
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
threadpoolctl==3.0.0
toml==0.10.2
typish==1.9.3
urllib3==1.26.7
vhash @ git+https://github.com/lakes-legendaries/vhash.git@8ee87458c35a0b1e6024916daefa5c7f2a8c6db0
xgboost==1.5.0
